Checkpoint Update
The checkpoints now have a billboard that sets the location where the player respawns, An enum that sets location (I want to use this for location names that appear on the screen but UI is wierd now so i dont know as tim to do them). You can now adjust the checkpoint size in the level. Todo is bugged so i can not vouch for if these checkpoints work